requirements.txt
fastapi
uvicorn
boto3
httpx
pydantic
python-dotenv
tavily-python
beautifulsoup4
